correction_matrix/matrix_multiplication: remove try ... catch
This function is not exposed to the user of the package.
Hence any errors we see here are an error in our code, and
overriding the error message is unnecessary.

If it is possible for the user to trigger a ValueError here
by providing bad inputs, validations must happen before
execution proceeds to here.