Add namedtuple to input, output. Cant remove parent.
Parent argument can't be removed from output as it is more than
just adding a parent column in the output (which has been removed
from final output). Structure of multiquant (which contains fragments)
is very different from maven input and output. Maven metabolite names
can't be equated with Component names in Multiquant which causes the
need of df_list in presence of parent argument and simple df otherwise