Remove parent argument
Parent argument was removed from some output function. The output
format of both maven and multiquant is consistent with each other
(as much as possible).Still certain functions have parent argument
which leads to user functions having this argument. Those cant be
avoided as labels of both mavena nd multiquant are defined differently.
Decorators need to implemented to take care of that.