Change sample name type validation to unicode
Sample Names are being validated as unicode rather than string.
The basic opertaions for both unicode and string are same. Unicode
is the format used for web applications to send data across. It is
safer to use. String type would raise an error while working with
webtool as data there is converted to unicode, hence the change.