Add function read_input_file
THis function takes path of the input file and covert it to pandas
data frame. First it checks for the extension of the file and then iot converts
the file using Pandas in built function. It also raises IOerror if the file
is is other than .xlsx,.xls,.txt,.csv