test: Write failing test corresponding to NCT-265
If for a given fragment, all samples are not present, the code
should give values 0 to intensities corresponding to the absent
samples. To find all samples for a fragment, take a union of
the samples in fragments dictionary and not just select samples
from first fragment. Some samples might be present in others but
missing in first fragment so the assumption that first fragment
contains all samples is incorrect.