Update element_data.json
R as an indistinguishable isotope is not being considered because it doesn't exist in the element dictionary. While considering derivatives in GCMS correction R is being used as a proxy for indistinguishable Carbons. For example in a compound with formula, C5H34O2R12, it is assumed that at max 5 Carbons can be labeled in the experiment, the rest 12 (given by R12) can't be labeled in the experiment. Lable for those 12 carbons can only come from indistinguishable isotopes.